requirements.txt
asttokens==2.4.1
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
certifi @ file:///C:/b/abs_8a944p1_gn/croot/certifi_1738623753421/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy @ file:///C:/b/abs_f2u2o_8s9g/croot/contourpy_1732540071787/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy==1.8.1
decorator==5.1.1
einops==0.8.1
executing==2.0.1
filelock @ file:///C:/b/abs_758187j281/croot/filelock_1744281404850/work
fonttools @ file:///C:/b/abs_4crkswws2h/croot/fonttools_1737040078745/work
fsspec==2025.3.2
ftfy==6.3.1
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
huggingface-hub==0.30.2
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
imageio @ file:///C:/b/abs_4cup9zfkvm/croot/imageio_1738160462075/work
imageio-ffmpeg==0.6.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver @ file:///C:/b/abs_faf90xet7a/croot/kiwisolver_1737040915779/work
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.10.0
matplotlib-inline==0.1.6
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mpmath_1699484863771/work
nest-asyncio==1.6.0
networkx @ file:///C:/b/abs_b054htfn9t/croot/networkx_1737043671910/work
numpy @ file:///C:/b/abs_71ejae7e2v/croot/numpy_and_numpy_base_1742570444043/work/dist/numpy-2.2.4-cp312-cp312-win_amd64.whl#sha256=51cd6ca08057acb115487cc25926841a5f12f8555a06e60179ffc632b9c70638
open_clip_torch==2.32.0
opencv-python==4.11.0.86
packaging==24.0
parso==0.8.3
pillow @ file:///C:/b/abs_68t826txdy/croot/pillow_1744613085333/work
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
pyparsing @ file:///C:/b/abs_40z8gyj9wi/croot/pyparsing_1731445739241/work
PyQt6==6.7.1
PyQt6_sip @ file:///C:/b/abs_f47pmp6xmn/croot/pyqt-split_1744804513582/work/pyqt_sip
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil==2.9.0.post0
pywin32==306
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
pyzmq==25.1.2
regex==2024.11.6
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
safetensors==0.5.3
scenedetect==0.6.6
setuptools==72.1.0
sip @ file:///C:/b/abs_5cto136kse/croot/sip_1738856220313/work
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
timm==1.0.15
tokenizers==0.21.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4
tqdm==4.67.1
traitlets==5.14.2
transformers==4.51.3
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
unicodedata2 @ file:///C:/b/abs_dfnftvxi4k/croot/unicodedata2_1736543771112/work
urllib3 @ file:///C:/b/abs_7bst06lizn/croot/urllib3_1737133657081/work
wcwidth==0.2.13
wheel==0.45.1
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
